[feature/APPC-982] - 1- Remove "ok_action" button from "activity_onboarding" and its listeners from OnboardingView and OnboardingPresenter; 2- Create separate files and rename the classes OnBoardingPageChangeListener, OnboardingPagerAdapter and OnboardingPageTransformer
