+++ app/src/main/java/com/asfoundation/wallet/ui/onboarding/OnboardingView.kt
@@ -10,8 +10,6 @@ interface OnboardingView {
 
   fun finishOnboarding()
 
-  fun getOkClick(): Observable<Any>
-
   fun getSkipClick(): Observable<Any>
 
   fun getCheckboxClick(): Observable<Any>
