+++ app/src/main/res/layout/activity_onboarding.xml
@@ -144,18 +144,5 @@
       app:layout_constraintTop_toBottomOf="@id/terms_conditions_layout"
       style="@style/ButtonStyle"
       />
-  <Button
-      android:id="@+id/ok_action"
-      android:layout_width="120dp"
-      android:layout_height="@dimen/button_height"
-      android:layout_marginBottom="20dp"
-      android:layout_marginTop="17dp"
-      android:text="@string/button_ok"
-      android:visibility="gone"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@id/terms_conditions_layout"
-      style="@style/ButtonStyle"
-      />
 </androidx.constraintlayout.widget.ConstraintLayout>
 </RelativeLayout>
